hashdb: show loading message and number of database items loaded

+++ hashdb.h
@@ -27,7 +27,7 @@ typedef struct _hashdb {
 extern hashdb_t *add_hashdb_entry(uint64_t path_hash, int pathlen, file_t *check);
 extern void dump_hashdb(hashdb_t *cur);
 extern int get_path_hash(char *path, uint64_t *path_hash);
-extern int load_hash_database(char *dbname);
+extern int64_t load_hash_database(char *dbname);
 extern int save_hash_database(const char * const restrict dbname);
 extern int read_hashdb_entry(file_t *file);
 extern int write_hashdb_entry(FILE *db, hashdb_t *cur, unsigned long *cnt);
